Implement some style consistencies
I only want to use arrow funcs when you have to put in an anonymous func
into some method. Defining funcs as arrow funcs in consts i think may
confuse the visual perception to see them as a random constant variable.
Also, the VScode outline view displays funcs better when they are
written as plain funcs. Another style consistency I will implement is
using the multiline-comment syntax for those comments that i put on
their own line to kind of be a header for a section of code. This is
primarily to differentiate them from commented-out code, and it just
looks like a title decoration.